docs(filter): update docs to use CustomCanvas
